JECC - Analysen - der "Play"-Knopf öffnet ein Dialog, welches ca. 10 mal geschlossen werden muss bis es weg ist fixes #938
